refactor(ums): move functions to separate module
